Add `npm run client` to Dockerfile.frontend
npm run client creates and populates src/generated. If npm run build is
run without running npm run client, it fails, as src/generated needs to
exist and be populated for all typescript files to be available.

The previous Dockerfile does not do this, which leads to a build failure.