Types (#6)
Added shared types that frontend and backend will use.
New intractable type for the Jukebox, new Song type that represents a song. Update, search, queue, initiate skip and vote for skip commands.
Queue uses url: String and not song: Song because the idea is that players will either queue an exact url that they know, or choose from searched songs that include url as part of Song.
Song has queuedBy?: Player in case we want to display which user queued a song ( '?' since we will have a default song/playlist).
Initiate skip command has player: Player so we can display which player started the vote and potentially ban that player from restarting a vote if it fails.